Configuration Reader (#18)
* Add boilerplate code

Adds the basic code for this feature, useful for a good start.

* Implement tests, and load real data

* Only read the defined env config

* Make host task read configuration

This is a very basic example that just uses a default configuration
file, it's not scoped yet. But it could be an idea, however I have no
idea on how the scopes work best.

Tried to use our logging function, it became clear that we need to read
the log level from somewhere global. Keept the old log statements in
comments for refence for now.

Refs: #1 #2